refactor: moved repository logic to backend/api/v1/repositories
- Relocated frontend repositories to backend repository layer
