+++ apps/backend/src/api/v1/repositories/sharedPageState.repository.ts
@@ -1,5 +1,5 @@
 // types
-import type Rom from '../types/Rom';
+import type Rom from '../../../../../../pokemon-collection/src/types/Rom';
 
 export type SharedPageState = {
   sharedMessage: string;
@@ -10,7 +10,7 @@ type Listener = () => void;
 
 class SharedPageStateRepository {
   private state: SharedPageState = {
-    sharedMessage: "",
+    sharedMessage: '',
     trackedRoms: [],
   };
