+++ apps/backend/src/api/v1/services/cardService.ts
@@ -1,5 +1,5 @@
 import type CardType from '../../../../../../pokemon-collection/src/types/Card';
-import { cardRepository } from '../../../../../../pokemon-collection/src/repositories/cardRepository';
+import { cardRepository } from '../repositories/cardRepository';
 
 class CardService {
   getCards(): CardType[] {
